benchmarks: wfi instead of busy loop in dummy threads
In the multicore case, we always need a second thread to run on the
second core (this is a bug; should be fixed).
Before, this second thread was just busy looping. This impacts the
performance of the first core due to RP2040's bus fabric, where
simultaneous instruction fetches aren't possible.
WFI on the second core solves this issue.