benchmarks/runqueue: test `RunQueue` with `N_CORES=CORES_NUMOF`
Before, we always tested (even if there was multicore support)
`RunQueue::<{ SCHED_PRIO_LEVELS }, { THREADS_NUMOF }>`.
This implies `N_CORES==1` because this is the default value for the
`N_CORES` Generic on `RunQueue`.

We actually have to test `N_CORES=`riot_rs_threads::CORES_NUMOF`
if we are on the multicore branch.